Refactor NodeMixin from BaseNode
